.claude/skills/fal-ai/scripts/requirements.txt
fal-client>=0.5.0
python-dotenv>=1.0.0
